move finalize rewards button to judge contest finding page
